Crate polyhal

Re-exports

pub use utils::addr::PhysAddr;

pub use utils::addr::VirtAddr;

pub use pagetable::MappingFlags;

pub use pagetable::MappingSize;

pub use pagetable::PageTable;

pub use pagetable::PageTableWrapper;

Modules

acpi

apic

common

consts

ctor

debug_console

This is a console for debugging, If you want to use this logging You need to use like this:

gdt

idt

instruction

Instruction Module.

irq

IRQ(Interrupt requests queue) module.

kcontext

Kernel Context module.

mem

multicore

Multi-core Module.

pagetable

percpu

Per-cpu module.

timer

Timer module.

utils

Macros

bit

bit macro will generate the number through a shift value.

bits

display_info

Display Platform Information with specified format display_info!(“item name”, “{}”, “format”); The output format like below: item name : format

pa

ph_ctor

Definiation a constructer

print

Print macro to print polyhal information with newline

println

Print macro to print polyhal information with newline

pub_use_arch

va

Constants

MEM_VECTOR_CAPACITY

Functions

hart_id

Attribute Macros

arch_entry

arch_interrupt

percpu