polyhal

Module debug_console

This is a console for debugging, If you want to use this logging You need to use like this:

Put a char to output device(always uart)

DebugConsole::putchar(b'3');

Get a char from input device(always uart)

DebugConsole::getchar();

Structs

DebugConsole

Functions

init_fb

Init the graphics console’s information, includes frame buffer addresse, width and height.