polyhal

Module multicore

Multi-core Module.

This is a leader for the multicore operation

You can use this function to use the multicore operation

Boot other calls after the multicore If you use this function call, you should call it after arch::init(..); This function will allocate the stack and map it for itself.

boot_core(hart_id, sp_top);

Here will have more functionality about multicore in the future.

Functions

boot_core