Struct polyhal::pagetable::PageTableWrapper

pub struct PageTableWrapper(pub PageTable);

Page Table Wrapper

You can use this wrapper to packing PageTable. If you release the PageTableWrapper, the PageTable will release its page table entry.

Tuple Fields
0: PageTable
Implementations

impl PageTableWrapper

Allocate a new PageTableWrapper with new page table root

This operation will restore the page table.

pub fn alloc() -> Self

Alloc a new PageTableWrapper with new page table root This operation will copy kernel page table space from booting page table.